+++ examples/serde_molecule_nostd/Cargo.toml
@@ -1,3 +1,7 @@
+[workspace]
+resolver = "2"
+members = []
+
 [package]
 name = "serde_molecule_nostd"
 version = "1.0.0"
@@ -8,3 +12,4 @@ edition = "2021"
 [dependencies]
 serde_molecule = { path = "../../serde_molecule", default-features = false, features = ["alloc"] }
 serde = { version = "1.0.208", default-features = false, features = ["derive"] }
+ckb-std = { version = "0.15.3", default-features = false, features = ["allocator", "ckb-types"] }
