+++ examples/serde_molecule_nostd/src/error.rs
@@ -1,0 +1,4 @@
+#[repr(i8)]
+pub enum Error {
+    Unknown = 30,
+}
