+++ serde_molecule/src/de.rs
@@ -4,7 +4,12 @@ use crate::{
     molecule::{disassemble_fixvec, disassemble_table, unpack_number},
     struct_serde::MoleculeStructDeserializer,
 };
+use alloc::vec;
+use alloc::vec::Vec;
 use serde::de::{self, value::U64Deserializer};
+
+#[cfg(not(feature = "std"))]
+use alloc::string::String;
 
 //////////////////////////////////////////////////////////////////////////////
 pub(crate) const DYNVEC_STR: &str = "__dyn_vec__";
