+++ serde_molecule/src/lib.rs
@@ -1,3 +1,4 @@
+#![cfg_attr(not(feature = "std"), no_std)]
 #![doc = include_str!("../../README.md")]
 
 #[cfg(not(any(feature = "std", feature = "alloc")))]
